refactor: Move selection hook into batch-actions view, reuse useExpansionToggle
Relocate the list-selection hook from the generic src/hooks/use-list-selection
into a dedicated batch-actions hook (useBatchActionSelection) that builds on the
existing useExpansionToggle, mirroring the useEventExpansionToggle precedent.

useExpansionToggle already models the "all vs. explicit set" state, so the new
hook reuses its isAllSelected/isSelected/toggleAll directly and layers on the two
batch-action specifics: individual items are locked while "select all" is active,
and selectedCount is reported from totalCount since the full item list isn't
available.
